% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d1_nm_poppk}
\alias{d1_nm_poppk}
\title{simulated theophylline data}
\description{
this dataset is an expanded version of the original theophylline population PK data.
Covariates \code{AGE} and \code{GENDER} were randomly assigned.
}
\author{
Yuan Xiong and Wenping Wang
}

