% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d2_nm_poppkpd}
\alias{d2_nm_poppkpd}
\title{simulated theophylline data}
\description{
this dataset was simulated by using theophylline population PK data and an indirect
response PD model.
Covariates \code{BMK1} and \code{BMK2} were randomly assigned.
}
\author{
Yuan Xiong and Wenping Wang
}

