% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d3_nm_idvpkpd}
\alias{d3_nm_idvpkpd}
\title{simulated theophylline data}
\description{
this dataset is similarly simulated as \code{d2_nm_poppkpd}, but with only a single subject.
}
\author{
Yuan Xiong and Wenping Wang
}

