% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instant.stan.extension.R
\name{instant.stan.extension}
\alias{instant.stan.extension}
\title{generate a customized NUTS-compatible ODE solver}
\usage{
instant.stan.extension(ode_str = NULL, covar = NULL, multi_state = FALSE)
}
\arguments{
\item{ode_str}{ODE string}

\item{covar}{a vector of character with names of covariates}

\item{multi_state}{a logical if multi-state}
}
\description{
generate a customized NUTS-compatible ODE solver
}
\details{
generate a customized NUTS-compatible ODE solver
}
\value{No return value, called for side effects}
\examples{
\donttest{
ode <- "
d/dt(depot) =-KA*depot;
d/dt(centr) = KA*depot - KE*centr;
"
instant.stan.extension(ode)
}
}
\author{
Wenping Wang
}

