\name{rstan-internal}
\alias{rstan-internal}
\alias{show,cxxdso-method}
\alias{RNG}
\alias{OUT}
\title{Internal Functions and Methods}
\description{
These functions are not intended to be called directly. In the case of
methods, they are documented elsewhere, either with the associated generic
function or with the class of the object for which the method is defined.
}
\value{No return value, called for side effects}
\keyword{internal}


